Move test credentials to `test/support/` dir
