--- auth.json
@@ -1,5 +1,0 @@
-{
-	"client_id": "1234",
-	"client_secret": "1234",
-	"refresh_token": "1234"
-}
