+++ test/support/testCredentials.json
@@ -1,0 +1,5 @@
+{
+	"client_id": "123456789",
+	"client_secret": "123456789",
+	"refresh_token": "123456789"
+}
